+++ cli/args.cpp
@@ -718,7 +718,7 @@ Args parseArguments(int argc, char** argv, unsigned &processed) {
         args.noColor = true;
         break;
       } else if (strEq(argv[a], "--version")) {
-        std::cout << "crate 0.8.17" << std::endl;
+        std::cout << "crate 0.8.18" << std::endl;
         exit(0);
       } else if (auto argShort = isShort(argv[a])) {
         switch (argShort) {
@@ -729,7 +729,7 @@ Args parseArguments(int argc, char** argv, unsigned &processed) {
           args.logProgress = true;
           break;
         case 'V':
-          std::cout << "crate 0.8.17" << std::endl;
+          std::cout << "crate 0.8.18" << std::endl;
           exit(0);
         default:
           err("unsupported short option '%s'", argv[a]);